@@ -1,43 +1,23 @@
 <template>
   <div id="app">
-    <a-layout>
-      <a-layout-header :style="{ position: 'fixed', zIndex: 1, width: '100%' }">
-        <Navbar />
-      </a-layout-header>
-      <a-layout-content>
-        <router-view :key="$route.fullPath" />
-      </a-layout-content>
-      <!-- <a-layout-footer>Footer</a-layout-footer> -->
-    </a-layout>
+    <HeaderComponent />
+    <main>
+      <!-- This is where your views will be displayed -->
+      <router-view></router-view>
+    </main>
+    <!-- <FooterComponent /> -->
   </div>
 </template>
 
 <script>
-import Navbar from "@/components/Navbar.vue";
+import HeaderComponent from '@/components/HeaderComponent';
+// import FooterComponent from '@/components/FooterComponent';
 
 export default {
+  name: 'App',
   components: {
-    Navbar,
+    HeaderComponent,
+    // FooterComponent,
   },
 };
 </script>
-
-<style>
-body {
-  margin: 0;
-  padding: 0;
-}
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-}
-.ant-layout {
-  background: white !important;
-}
-.ant-layout-content {
-  margin-top: 64px;
-}
-</style>
